+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2023.1.7
 
 * Support browser localStorage
+* Support local json files
 
 ## 2023.1.6